migrated to component-based architecture, shared reset flow
