+++ frontend/src/app/reset-password/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ForgotPassword from "../../components/ForgotPassword";
+
+export default function ResetPasswordPage() {
+  return <ForgotPassword initialStage="verify" />;
+}
